+++ src/compress/zstd_double_fast.rs
@@ -452,8 +452,8 @@ unsafe extern "C" fn ZSTD_hash4Ptr(
 }
 static mut prime5bytes: u64 = 889523592379;
 unsafe extern "C" fn ZSTD_hash5(mut u: u64, mut h: u32, mut s: u64) -> usize {
-    return ((u << 64 - 40 as std::ffi::c_int) * prime5bytes ^ s)
-        >> 64_u32.wrapping_sub(h);
+    return (((u << 64 - 40 as std::ffi::c_int) * prime5bytes ^ s)
+        >> 64_u32.wrapping_sub(h)) as usize;
 }
 unsafe extern "C" fn ZSTD_hash5Ptr(
     mut p: *const std::ffi::c_void,
@@ -463,8 +463,8 @@ unsafe extern "C" fn ZSTD_hash5Ptr(
 }
 static mut prime6bytes: u64 = 227718039650203;
 unsafe extern "C" fn ZSTD_hash6(mut u: u64, mut h: u32, mut s: u64) -> usize {
-    return ((u << 64 - 48 as std::ffi::c_int) * prime6bytes ^ s)
-        >> 64_u32.wrapping_sub(h);
+    return (((u << 64 - 48 as std::ffi::c_int) * prime6bytes ^ s)
+        >> 64_u32.wrapping_sub(h)) as usize;
 }
 unsafe extern "C" fn ZSTD_hash6Ptr(
     mut p: *const std::ffi::c_void,
@@ -474,8 +474,8 @@ unsafe extern "C" fn ZSTD_hash6Ptr(
 }
 static mut prime7bytes: u64 = 58295818150454627;
 unsafe extern "C" fn ZSTD_hash7(mut u: u64, mut h: u32, mut s: u64) -> usize {
-    return ((u << 64 - 56 as std::ffi::c_int) * prime7bytes ^ s)
-        >> 64_u32.wrapping_sub(h);
+    return (((u << 64 - 56 as std::ffi::c_int) * prime7bytes ^ s)
+        >> 64_u32.wrapping_sub(h)) as usize;
 }
 unsafe extern "C" fn ZSTD_hash7Ptr(
     mut p: *const std::ffi::c_void,
@@ -485,7 +485,7 @@ unsafe extern "C" fn ZSTD_hash7Ptr(
 }
 static mut prime8bytes: u64 = 0xcf1bbcdcb7a56463 as std::ffi::c_ulonglong as u64;
 unsafe extern "C" fn ZSTD_hash8(mut u: u64, mut h: u32, mut s: u64) -> usize {
-    return (u * prime8bytes ^ s) >> 64_u32.wrapping_sub(h);
+    return ((u * prime8bytes ^ s) >> 64_u32.wrapping_sub(h)) as usize;
 }
 unsafe extern "C" fn ZSTD_hash8Ptr(
     mut p: *const std::ffi::c_void,
@@ -655,12 +655,12 @@ unsafe extern "C" fn ZSTD_countLeadingZeros64(mut val: u64) -> std::ffi::c_uint 
 unsafe extern "C" fn ZSTD_NbCommonBytes(mut val: usize) -> std::ffi::c_uint {
     if MEM_isLittleEndian() != 0 {
         if MEM_64bits() != 0 {
-            return ZSTD_countTrailingZeros64(val) >> 3
+            return ZSTD_countTrailingZeros64(val as u64) >> 3
         } else {
             return ZSTD_countTrailingZeros32(val as u32) >> 3
         }
     } else if MEM_64bits() != 0 {
-        return ZSTD_countLeadingZeros64(val) >> 3
+        return ZSTD_countLeadingZeros64(val as u64) >> 3
     } else {
         return ZSTD_countLeadingZeros32(val as u32) >> 3
     };
